# api.yaml
openapi: 3.0.3
info:
  title: API Системы учета и выдачи инструментов
  description: |
    Спецификация REST API для синхронизации, авторизации и управления реестром оборудования с проверкой Ed25519 подписей оффлайн.
    Подробная логика — в документе «Описание проекта подробно, используемое в разработке».

    ## Время
    Все метки времени в API и в протоколе QR — Unix time в секундах (UTC, int64); от часового пояса не зависят.
    Клиенты работают по времени сервера: `serverTime` из `/api/auth/server-key` и `/api/auth/worker-cert`
    задаёт смещение локальных часов.

    ## Авторизация
    JWT передаётся в заголовке `Authorization: Bearer <accessToken>`. Срок действия — 12 часов.
    Содержит id пользователя (claim `sub`), роль (claim `role`) и id его компании (claim `company_id`);
    подписан HMAC-SHA256 (секрет `JWT_SIGNING_KEY`).
    Отзыва нет: после увольнения выданный JWT действует до истечения срока.

    ## Идентификаторы
    id пользователей, компаний и инструментов — UUID в каноническом виде (строчные hex, `8-4-4-4-12`). Их нельзя
    перебрать и по ним нельзя оценить число пользователей или инструментов. Защитой данных они не являются:
    от доступа к чужому объекту защищает проверка компании (см. ниже), а не секретность id.
    Значение, не являющееся UUID, в пути или запросе — 400.
    id записей журнала (`logId`, `RentalLogEntry.id`) — обычные числа: на вход они нигде не принимаются,
    записи всегда выбираются по компании.

    ## Компании
...

...

...

...

...

...
servers:
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      operationId: GetUsers
      summary: Список пользователей своей компании
      description: "**Требуется роль:** Owner. Только пользователи компании вызывающего."
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список пользователей своей компании (включая неподтверждённых и уволенных)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    delete:
      tags: [Users]
      operationId: DeactivateUser
      summary: Уволить/Деактивировать пользователя
      description: |
        Переводит is_active = false; запись не удаляется (на неё ссылается журнал). Повторный вызов — 200.
        Уже выданные JWT и сертификаты не отзываются и действуют до истечения срока (не более 12 часов).
        Owner не может уволить сам себя. Только пользователь своей компании: пользователь другой компании — 404.

        **Требуется роль:** Owner.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Пользователь уволен (is_active = false), без тела
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Попытка уволить самого себя
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://tools.ietf.org/html/rfc9110#section-15.5.10
                title: Conflict
                status: 409
                detail: Нельзя уволить самого себя
                traceId: 00-2f1c...-01
  /api/users/{id}/approve:
    post:
      tags: [Users]
      operationId: ApproveUser
      summary: Подтверждение регистрации пользователя
      description: |
        Переводит is_approved = true. Повторный вызов — 200.
        Только пользователь своей компании: пользователь другой компании — 404.

        **Требуется роль:** Owner.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Пользователь подтверждён, без тела
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # УПРАВЛЕНИЕ ИНСТРУМЕНТАМИ
  # ==========================================
  /api/tools:
    get:
      tags: [Tools]
      operationId: GetTools
      summary: Получение реестра инструментов
      description: |
        Возвращает список активных (не списанных) инструментов своей компании вместе с текущим держателем —
        локальный реестр завхоза и печать QR-кодов.

        **Требуется роль:** Issuer, Owner.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список инструментов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Tools]
      operationId: CreateTool
      summary: Добавление нового инструмента
      description: |
        Инструмент добавляется в реестр компании вызывающего. id присваивается базой данных и возвращается
        в ответе — по нему завхоз печатает QR-код инструмента.
        Новый инструмент создаётся в состоянии Good, без комментария и без держателя.
        Название — не длиннее 100 символов (попадает в QR сотрудника).

        **Требуется роль:** Issuer, Owner.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
      responses:
        '201':
          description: Инструмент создан, ему присвоен id
          headers:
            Location:
              description: Адрес созданного инструмента, `/api/tools/{id}`
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/tools/{id}:
    get:
      tags: [Tools]
      operationId: GetTool
      summary: Данные инструмента
      description: |
        Возвращает инструмент своей компании, в том числе списанный. Инструмент другой компании — 404.

        **Требуется роль:** Issuer, Owner.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Информация об инструменте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Tools]
      operationId: UpdateToolCondition
      summary: Частичное обновление состояния инструмента
      description: |
        Ручное вмешательство вне цикла выдачи/возврата (заметили повреждение, инструмент отремонтирован).
        Обычный путь обновления `condition` — VALID-запись в `/api/sync`.

        Поле, которое не передано (или null), не меняется; `comment: ""` очищает комментарий.
        Изменение `condition` выставляет время изменения состояния в текущее время сервера — после этого
        более ранние операции из поздних синхронизаций состояние не перезаписывают.
        Разрешено и для списанного инструмента (is_active не меняется).
        Инструмент другой компании — 404.

        **Требуется роль:** Issuer, Owner.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateToolConditionRequest'
      responses:
        '200':
          description: Состояние успешно обновлено, без тела
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tools]
      operationId: DeleteTool
      summary: Списание инструмента
      description: |
        Переводит is_active = false. Необратимое действие. Повторный вызов — 200.
        Списанный инструмент не попадает в GET /api/tools, но остаётся в GET /api/tools/{id} и в журнале.
        Инструмент другой компании — 404.

        **Требуется роль:** Owner.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Инструмент списан, без тела
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==========================================
  # СИНХРОНИЗАЦИЯ И ЖУРНАЛ (LOGS)
  # ==========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT из POST /api/auth/login, срок действия 12 часов
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: id объекта (UUID); значение, не являющееся UUID, — 400
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: |
        Некорректный запрос: тело не разбирается как JSON нужной формы, нарушены обязательность или ограничения полей,
        неверный параметр пути или запроса. Ключи `errors` — пути к полям в именах JSON.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc9110#section-15.5.1
            title: One or more validation errors occurred.
            status: 400
            errors:
              fullName: ['The field FullName must be a string or array type with a maximum length of ''100''.']
              role: ['The Role field is required.']
            traceId: 00-2f1c...-01
    Unauthorized:
      description: Не передан, просрочен или недействителен JWT. Без тела; заголовок `WWW-Authenticate` со схемой Bearer
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer error="invalid_token", error_description="The token expired at '...'"
    Forbidden:
      description: |
        Роль пользователя не допускается к этому эндпоинту. Без тела.
        Увольнение уже выданный JWT не отзывает (см. раздел «Авторизация»).
    NotFound:
      description: Объект с таким id не найден (в том числе принадлежащий другой компании)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://tools.ietf.org/html/rfc9110#section-15.5.5
            title: Not Found
            status: 404
            detail: Инструмент с id 17c4e5a2-93b8-4f10-8d6a-0b2e7c9d4f35 не найден
            traceId: 00-2f1c...-01
  schemas:
...
    ProblemDetails:
      type: object
      description: |
        Описание ошибки в формате RFC 9457 (ранее RFC 7807), `application/problem+json`.
        Формирует ASP.NET Core; допускает дополнительные поля.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: URI типа ошибки (для стандартных кодов — ссылка на раздел RFC 9110)
          example: https://tools.ietf.org/html/rfc9110#section-15.5.5
        title:
          type: string
          description: Краткое название типа ошибки (стандартное, на английском)
          example: Not Found
        status:
          type: integer
          format: int32
          description: HTTP-код ответа
          example: 404
        detail:
          type: string
          nullable: true
          description: Пояснение к конкретному случаю на русском, для показа пользователю
          example: Инструмент с id 17c4e5a2-93b8-4f10-8d6a-0b2e7c9d4f35 не найден
        instance:
          type: string
          nullable: true
          description: URI конкретного случая ошибки (обычно не заполняется)
        traceId:
          type: string
          description: Идентификатор запроса для поиска в логах сервера
          example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-00
      additionalProperties: true
    ValidationProblemDetails:
      description: ProblemDetails с перечнем ошибок по полям (ответ 400)
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          required: [errors]
          properties:
            errors:
              type: object
              description: |
                Ключ — путь к полю в именах JSON (`fullName`, `[3].qr`, `id`) или JSON-путь при ошибке разбора
                (`$.role`); значение — список сообщений об ошибках.
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                '[3].qr': ['Строка не соответствует формату AT1: ожидается 5 частей через точку']

    # ---------- Перечисления ----------
    UserRole:
      type: string
      description: Worker — сотрудник, Issuer — завхоз, Owner — владелец (регистрируются все три)
      enum: [Worker, Issuer, Owner]
    ToolCondition:
      type: string
      description: Состояние инструмента, по возрастанию тяжести (Good < Damaged < Broken)
      enum: [Good, Damaged, Broken]
...
    CreateToolRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
          description: Название. Попадает в QR сотрудника (toolName в op), поэтому ограничено 100 символами
          example: "Дрель Bosch GSB 16 RE"
    UpdateToolConditionRequest:
      type: object
      description: Не переданное (или null) поле не меняется; comment = "" очищает комментарий.
      properties:
        condition:
          # В OpenAPI 3.0 nullable действует только рядом с type, поэтому значения ToolCondition перечислены здесь
          type: string
          enum: [Good, Damaged, Broken]
          nullable: true
          description: Новое состояние; меняет и время изменения состояния (текущее время сервера)
        comment:
          type: string
          nullable: true
          description: Новый комментарий; "" — очистить
      example:
        condition: Damaged
        comment: Трещина на корпусе
...
    UserResponse:
      type: object
      required: [id, fullName, phone, role, isApproved, isActive]
      properties:
        id:
          type: string
          format: uuid
          example: 5b1d0c8a-2f4e-4a63-9d17-8e0c3a6f4b21
        fullName:
          type: string
          example: "Иванов Иван Иванович"
        phone:
          type: string
          pattern: '^\+7[0-9]{10}$'
          description: Нормализованный номер
          example: "+79991234567"
        role:
          $ref: '#/components/schemas/UserRole'
        isApproved:
          type: boolean
          description: Регистрация подтверждена Owner компании (Owner, создавший компанию, подтверждён сразу)
        isActive:
          type: boolean
          description: false — уволен (запись не удаляется)
    Tool:
      type: object
      description: |
        Держатель (holderId, holderName, heldSince) вычисляется сервером по последней по времени операции
        VALID-записи журнала (view tool_current_holders): если это TAKE — инструмент на руках у сотрудника,
        если GIVE или записей нет — на месте у завхоза, и все три поля равны null.
        Учитываются только синхронизированные записи. Инструмент принадлежит компании вызывающего (поля компании в модели нет).
      required: [id, name, condition, comment, isActive, holderId, holderName, heldSince]
      properties:
        id:
          type: string
          format: uuid
          description: Присваивается базой данных; по нему печатается QR инструмента
          example: 17c4e5a2-93b8-4f10-8d6a-0b2e7c9d4f35
        name:
          type: string
          example: "Дрель Bosch GSB 16 RE"
        condition:
          $ref: '#/components/schemas/ToolCondition'
        comment:
          type: string
          nullable: true
          description: Комментарий (ставится только через PATCH)
          example: null
        isActive:
          type: boolean
          description: false — списан (необратимо)
        holderId:
          type: string
          format: uuid
          nullable: true
          description: id сотрудника, у которого инструмент на руках; null — инструмент на месте у завхоза
          example: 5b1d0c8a-2f4e-4a63-9d17-8e0c3a6f4b21
        holderName:
          type: string
          nullable: true
          description: ФИО сотрудника-держателя (из users)
          example: "Иванов Иван Иванович"
        heldSince:
          type: integer
          format: int64
          nullable: true
          description: Время операции TAKE, с которой инструмент у держателя (Unix time)
          example: 1790003600

    # ---------- Содержимое QR сотрудника ----------
...