# backend/AccountThis.Api/Models/UserModels.cs
namespace AccountThis.Api.Models;

// components.schemas.UserResponse
public class UserResponse
{
    public Guid Id { get; set; }

    public string FullName { get; set; } = string.Empty;

    // Нормализованный: +7XXXXXXXXXX
    public string Phone { get; set; } = string.Empty;

    public UserRole Role { get; set; }

    public bool IsApproved { get; set; }

    public bool IsActive { get; set; }
}